remove users blogs and comments when user deleted

+++ includes/functions.php
@@ -202,6 +202,26 @@ function editComment($conn, $uuid, $blog_uuid, $description)
 
 function deleteUser($conn, $uuid)
 {
+    $sql = "DELETE FROM blogs WHERE user_uuid = ?";
+    $stmt = mysqli_stmt_init($conn);
+    if (!mysqli_stmt_prepare($stmt, $sql)) {
+        header("location: ../profiles/$uuid.php?error=stmtfailed");
+        exit();
+    }
+    mysqli_stmt_bind_param($stmt, "s", $uuid);
+    mysqli_stmt_execute($stmt);
+    mysqli_stmt_close($stmt);
+
+    $sql = "DELETE FROM comments WHERE user_uuid = ?";
+    $stmt = mysqli_stmt_init($conn);
+    if (!mysqli_stmt_prepare($stmt, $sql)) {
+        header("location: ../profiles/$uuid.php?error=stmtfailed");
+        exit();
+    }
+    mysqli_stmt_bind_param($stmt, "s", $uuid);
+    mysqli_stmt_execute($stmt);
+    mysqli_stmt_close($stmt);
+
     $sql = "DELETE FROM users WHERE uuid = ?";
     $stmt = mysqli_stmt_init($conn);
     if (!mysqli_stmt_prepare($stmt, $sql)) {
